+++ src/Graphify.Tests/Snippets/Declarations/Boilerplate.cs
@@ -2,11 +2,6 @@
 
 public static class Boilerplate
 {
-    public static readonly Generated Base = new(
-        NavigatorGenerator.Content,
-        NavigatorGenerator.Hint,
-        typeof(NavigatorGenerator));
-
     public static readonly Generated Graphify = new(
         GraphifyAttributeGenerator.Content,
         $"{GraphifyAttributeGenerator.Name}Attribute.g.cs",
@@ -17,10 +12,10 @@ public static class Boilerplate
         $"{TraverseAttributeGenerator.Name}Attribute.g.cs",
         typeof(TraverseAttributeGenerator));
 
-    public static readonly Generated Navigator = new(
-        NavigatorContractGenerator.Content,
-        NavigatorContractGenerator.Hint,
-        typeof(NavigatorContractGenerator));
+    public static readonly Generated Inspector = new(
+        InspectorContractGenerator.Content,
+        InspectorContractGenerator.Hint,
+        typeof(InspectorContractGenerator));
 
     public static readonly Generated Visitor = new(
         VisitorContractGenerator.Content,